+++ src/Application/Audit/AuditAction.php
@@ -37,6 +37,7 @@ enum AuditAction: string
     case SEATING_SEAT_UPDATED = 'seating.seat_updated';
     case SEATING_GROUP_CREATED = 'seating.group_created';
     case SEATING_GROUP_UPDATED = 'seating.group_updated';
+    case SEATING_GROUP_MOVED = 'seating.group_moved';
     case SEATING_RECOMMENDATION_GENERATED = 'seating.recommendation_generated';
     case SEATING_RECOMMENDATION_APPLIED = 'seating.recommendation_applied';
     case SEATING_ASSIGNED = 'seating.assigned';